Revert template changes
