+++ packages/commerce-sdk-react/src/provider.tsx
@@ -5,11 +5,25 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import React, {ReactElement, useEffect, useMemo} from 'react'
-import {ShopperBasketsTypes} from 'commerce-sdk-isomorphic'
 import Auth from './auth'
 import {ApiClientConfigParams, ApiClients} from './hooks/types'
 import {Logger} from './types'
-import {DWSID_COOKIE_NAME, SERVER_AFFINITY_HEADER_KEY} from './constant'
+import {DWSID_COOKIE_NAME, MOBIFY_PATH, SERVER_AFFINITY_HEADER_KEY, SLAS_PRIVATE_PROXY_PATH} from './constant'
+import {
+    ShopperBaskets,
+    ShopperContexts,
+    ShopperCustomers,
+    ShopperExperience,
+    ShopperLogin,
+    ShopperOrders,
+    ShopperProducts,
+    ShopperPromotions,
+    ShopperGiftCertificates,
+    ShopperSearch,
+    ShopperSeo,
+    ShopperBasketsTypes,
+    ShopperStores
+} from 'commerce-sdk-isomorphic'
 
 export interface CommerceApiProviderProps extends ApiClientConfigParams {
     children: React.ReactNode
@@ -36,7 +50,11 @@ export interface CommerceApiProviderProps extends ApiClientConfigParams {
 }
 
 export type ParameterFilter<T> = (params: T, methodName: string) => Partial<T>
-export type ParameterTransformer<T> = (params: T, methodName: string, options: any) => any | Promise<any>
+export type ParameterTransformer<T> = (
+    params: T,
+    methodName: string,
+    options: any
+) => any | Promise<any>
 export type BeforeCallCallback<TParams> = (
     methodName: string,
     params: TParams,
@@ -226,46 +244,101 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         serverAffinityHeader[SERVER_AFFINITY_HEADER_KEY] = dwsid
     }
 
-    const defaultTransformer: ParameterTransformer<Record<string, any>> = (_, _$, options) => {
-        return {
-            ...options,
-            headers: {
-                ...options.headers,
-                ...serverAffinityHeader
-            },
-            throwOnBadResponse: true,
-            fetchOptions: {
-                ...options.fetchOptions,
-                ...fetchOptions
+    let updatedClients: ApiClients
+
+    if (apiClients) {
+        const defaultTransformer: ParameterTransformer<Record<string, any>> = (_, _$, options) => {
+            return {
+                ...options,
+                headers: {
+                    ...options.headers,
+                    ...serverAffinityHeader
+                },
+                throwOnBadResponse: true,
+                fetchOptions: {
+                    ...options.fetchOptions,
+                    ...fetchOptions
+                }
             }
         }
-    }
 
-    const updatedClients = useMemo(() => {
-        const clients: Record<string, any> = {}
+        updatedClients = useMemo(() => {
+            const clients: Record<string, any> = {}
 
-        Object.entries(apiClients ?? {}).forEach(([key, apiClient]) => {
-            clients[key] = withParameterInjection(apiClient, {
-                props,
-                transformer: transformer ?? defaultTransformer,
-                onBeforeCall,
-                onAfterCall,
-                onError
+            Object.entries(apiClients ?? {}).forEach(([key, apiClient]) => {
+                clients[key] = withParameterInjection(apiClient, {
+                    props,
+                    transformer: transformer ?? defaultTransformer,
+                    onBeforeCall,
+                    onAfterCall,
+                    onError
+                })
             })
-        })
 
-        return clients as ApiClients
-    }, [
-        clientId,
-        organizationId,
-        shortCode,
-        siteId,
-        proxy,
-        fetchOptions,
-        locale,
-        currency,
-        headers?.['correlation-id']
-    ])
+            return clients as ApiClients
+        }, [
+            clientId,
+            organizationId,
+            shortCode,
+            siteId,
+            proxy,
+            fetchOptions,
+            locale,
+            currency,
+            headers?.['correlation-id']
+        ])
+    } else {
+        const config = {
+            proxy,
+            headers: {
+                ...headers,
+                ...serverAffinityHeader
+            },
+            parameters: {
+                clientId,
+                organizationId,
+                shortCode,
+                siteId,
+                locale,
+                currency
+            },
+            throwOnBadResponse: true,
+            fetchOptions
+        }
+
+        const baseUrl = config.proxy.split(MOBIFY_PATH)[0]
+        const privateClientEndpoint = `${baseUrl}${SLAS_PRIVATE_PROXY_PATH}`
+
+        updatedClients = useMemo(() => {
+            return {
+                shopperBaskets: new ShopperBaskets(config),
+                shopperContexts: new ShopperContexts(config),
+                shopperCustomers: new ShopperCustomers(config),
+                shopperExperience: new ShopperExperience(config),
+                shopperGiftCertificates: new ShopperGiftCertificates(config),
+                shopperLogin: new ShopperLogin({
+                    ...config,
+                    proxy: enablePWAKitPrivateClient ? privateClientEndpoint : config.proxy
+                }),
+                shopperOrders: new ShopperOrders(config),
+                shopperProducts: new ShopperProducts(config),
+                shopperPromotions: new ShopperPromotions(config),
+                shopperSearch: new ShopperSearch(config),
+                shopperSeo: new ShopperSeo(config),
+                shopperStores: new ShopperStores(config)
+            }
+        }, [
+            clientId,
+            organizationId,
+            shortCode,
+            siteId,
+            proxy,
+            fetchOptions,
+            locale,
+            currency,
+            headers?.['correlation-id']
+        ])
+    }
 
     // Initialize the session
     useEffect(() => void auth.ready(), [auth])
