+++ packages/template-retail-react-app/package.json
@@ -60,7 +60,6 @@
     "base64-arraybuffer": "^0.2.0",
     "bundlesize2": "^0.0.35",
     "card-validator": "^8.1.1",
-    "commerce-sdk-isomorphic": "^3.3.0",
     "cross-env": "^5.2.1",
     "cross-fetch": "^3.1.4",
     "focus-visible": "^5.2.0",
